Move queue classes to subnamespace
